+++ packages/skirmish/src/Healthbar.res.mjs
@@ -15,14 +15,14 @@ let middle = GameContext$Skirmish.k.Color.fromHex("#ffdf20");
 let bad = GameContext$Skirmish.k.Color.fromHex("#e7000b");
 
 function getHealthColor(healthPercent) {
-  if (healthPercent >= 70) {
-    let t = (100 - healthPercent) / 30;
+  if (healthPercent >= 60) {
+    let t = (100 - healthPercent) / (100 - 60);
     return good.lerp(middle, t);
   }
-  if (healthPercent < 40) {
+  if (healthPercent < 20) {
     return bad;
   }
-  let t$1 = (70 - healthPercent) / 30;
+  let t$1 = (60 - healthPercent) / (60 - 20);
   return middle.lerp(bad, t$1);
 }
 
@@ -122,7 +122,7 @@ function make(pokemon) {
       id: "healthbar",
       draw: draw
     },
-    pokemon.team === false ? GameContext$Skirmish.k.pos(10, 10) : GameContext$Skirmish.k.pos(GameContext$Skirmish.k.width() - 160, GameContext$Skirmish.k.height() - 50)
+    pokemon.team === false ? GameContext$Skirmish.k.pos(10, 10) : GameContext$Skirmish.k.pos(GameContext$Skirmish.k.width() / 2 - 10, GameContext$Skirmish.k.height() - 50)
   ]);
   pokemon.onHurt(_deltaHP => {
     let newHealthPercent = Pokemon$Skirmish.getHealthPercentage(pokemon);
@@ -131,10 +131,16 @@ function make(pokemon) {
   return healthbar;
 }
 
+let middleUpperLimit = 60;
+
+let middleLowerLimit = 20;
+
 export {
   good,
   middle,
   bad,
+  middleUpperLimit,
+  middleLowerLimit,
   getHealthColor,
   setHealth,
   draw,